+++ Source/PropertyTools.Wpf.Tests/Extensions/ReflectionExtensionsTests.cs
@@ -7,6 +7,7 @@
 namespace PropertyTools.Wpf.Tests
 {
     using System;
+    using System.Collections.ObjectModel;
     using System.ComponentModel;
     using System.Linq;
 
@@ -28,11 +29,58 @@ namespace PropertyTools.Wpf.Tests
             Value3
         }
 
+        [AttributeUsage(AttributeTargets.Property)]
+        private class TestBaseAttribute : Attribute
+        {
+        }
+
+        private class TestDerivedAttribute : TestBaseAttribute
+        {
+        }
+
+        private class ReflectionTestModel
+        {
+            [TestDerivedAttribute]
+            public ObservableCollection<string> Values { get; } = new ObservableCollection<string>();
+
+            public string Name { get; set; }
+        }
+
         [Test]
         public void FilterOnBrowsableAttribute()
         {
             Assert.That(Enum.GetValues(typeof(Enum1)).Length, Is.EqualTo(5));
             Assert.That(Enum.GetValues(typeof(Enum1)).FilterOnBrowsableAttribute().Count(), Is.EqualTo(3));
+        }
+
+        [Test]
+        public void GetFirstAttributeOrDefault_ExactAttributeType_ReturnsAttribute()
+        {
+            var descriptor = TypeDescriptor.GetProperties(typeof(ReflectionTestModel))[nameof(ReflectionTestModel.Values)];
+
+            var attribute = descriptor.GetFirstAttributeOrDefault(typeof(TestDerivedAttribute));
+
+            Assert.That(attribute, Is.TypeOf<TestDerivedAttribute>());
+        }
+
+        [Test]
+        public void GetFirstAttributeOrDefault_DerivedAttribute_ReturnsAttribute()
+        {
+            var descriptor = TypeDescriptor.GetProperties(typeof(ReflectionTestModel))[nameof(ReflectionTestModel.Values)];
+
+            var attribute = descriptor.GetFirstAttributeOrDefault(typeof(TestBaseAttribute));
+
+            Assert.That(attribute, Is.TypeOf<TestDerivedAttribute>());
+        }
+
+        [Test]
+        public void GetFirstAttributeOrDefault_MissingAttribute_ReturnsNull()
+        {
+            var descriptor = TypeDescriptor.GetProperties(typeof(ReflectionTestModel))[nameof(ReflectionTestModel.Name)];
+
+            var attribute = descriptor.GetFirstAttributeOrDefault(typeof(TestBaseAttribute));
+
+            Assert.That(attribute, Is.Null);
         }
     }
 }